requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
